Muchas cosas

+++ src/java/proyectoTAW/servlet/EliminarCategoriaServlet.java
@@ -12,8 +12,6 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import proyectoTAW.dto.CategoriaDTO;
-import proyectoTAW.entity.Categoria;
 import proyectoTAW.service.CategoriaService;
 
 /**
@@ -21,7 +19,7 @@ import proyectoTAW.service.CategoriaService;
  * @author juanm
  */
 @WebServlet(name = "EliminarCategoriaServlet", urlPatterns = {"/EliminarCategoriaServlet"})
-public class EliminarCategoriaServlet extends HttpServlet {
+public class EliminarCategoriaServlet extends ProjectoTAWServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -37,12 +35,13 @@ public class EliminarCategoriaServlet extends HttpServlet {
     
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        
+        if(super.redirigirUsuario(request, response, "Administrador", request.getSession())){
         String str = request.getParameter("id");
   
         this.cService.remove(Integer.parseInt(str));
         
         response.sendRedirect(request.getContextPath() + "/EditorCategoriasServlet");
+        }
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
