+++ src/java/proyectoTAW/servlet/EliminarUsuarioServlet.java
@@ -6,15 +6,12 @@
 package proyectoTAW.servlet;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import javax.ejb.EJB;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import proyectoTAW.dao.UsuarioFacade;
-import proyectoTAW.entity.Usuario;
 import proyectoTAW.service.UsuarioService;
 
 /**
@@ -22,7 +19,7 @@ import proyectoTAW.service.UsuarioService;
  * @author juanm
  */
 @WebServlet(name = "EliminarUsuarioServlet", urlPatterns = {"/EliminarUsuarioServlet"})
-public class EliminarUsuarioServlet extends HttpServlet {
+public class EliminarUsuarioServlet extends ProjectoTAWServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -38,12 +35,14 @@ public class EliminarUsuarioServlet extends HttpServlet {
     
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        if(super.redirigirUsuario(request, response, "Administrador", request.getSession())){
         response.setContentType("text/html;charset=UTF-8");
         
         String str = request.getParameter("id");
         this.uService.remove(Integer.parseInt(str));
         
         response.sendRedirect(request.getContextPath() + "/ListaUsuariosServlet?filtro=1");
+        }
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
